workspace/requirements.txt
albumentations==1.3.0
face_recognition==1.3.0
gradio==3.32.0
hydra-core==1.3.2
matplotlib==3.5.2
numpy==1.24.2
omegaconf==2.3.0
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
pandas==1.4.4
Pillow==9.5.0
pyrootutils==1.0.4
torch==1.13.1
torch_directml==0.1.13.1.dev230301
torchvision==0.14.1
